Deprecate empty sources field values on generators.
[ci skip-rust]
[ci skip-build-wheels]
